feat: rename aapi package
