+++ apps/aants/src/helpers/notificationDispatch.ts
@@ -1,9 +1,8 @@
-import { WebsocSection } from '@packages/anteater-api-types';
+import { WebsocSection } from '@packages/anteater-api/types';
 import { render, toPlainText } from '@react-email/render';
 
 import { CourseNotificationEmail } from '../emails/CourseNotificationEmail';
 import { aantsEnvSchema } from '../env';
-
 import { queueEmail } from './emailQueue';
 import { User } from './subscriptionData';
