+++ apps/antalmanac/package.json
@@ -37,7 +37,7 @@
         "@mui/system": "^7.3.7",
         "@mui/x-date-pickers": "^8.3.1",
         "@packages/antalmanac-types": "workspace:^",
-        "@packages/anteater-api-types": "workspace:^",
+        "@packages/anteater-api": "workspace:^",
         "@paralleldrive/cuid2": "^2.2.2",
         "@reactour/tour": "^3.6.1",
         "@tanstack/react-query": "^5.0.0",
