--- packages/anteater-api-types/.gitignore
@@ -1,1 +1,0 @@
-src/generated/
