--- packages/anteater-api-types/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@packages/anteater-api-types",
-  "version": "0.0.1",
-  "description": "Anteater API types for AntAlmanac",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "type": "module",
-  "scripts": {
-      "postinstall": "openapi-typescript https://anteaterapi.com/openapi.json -o src/generated/anteater-api-types.ts"
-  },
-  "devDependencies": {
-    "typescript": "^5.7.2",
-      "openapi-typescript": "7.4.3"
-  }
-}
