+++ packages/anteater-api/.gitignore
@@ -1,0 +1,1 @@
+src/types/generated/
