+++ packages/anteater-api/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "@packages/anteater-api",
+    "version": "0.0.1",
+    "description": "Anteater API types, client, and utilities for AntAlmanac",
+    "type": "module",
+    "exports": {
+        "./types": "./src/types/index.ts",
+        "./client": "./src/client/index.ts",
+        "./utils": "./src/utils/index.ts"
+    },
+    "scripts": {
+        "postinstall": "openapi-typescript https://anteaterapi.com/openapi.json -o src/types/generated/anteater-api-types.ts"
+    },
+    "devDependencies": {
+        "typescript": "^5.7.2",
+        "openapi-typescript": "7.4.3"
+    }
+}
