+++ packages/anteater-api/src/types/courses.ts
