+++ packages/types/src/courseData.ts
@@ -1,4 +1,4 @@
-import { WebsocSection, WebsocSectionType } from '@packages/anteater-api-types';
+import { WebsocSection, WebsocSectionType } from '@packages/anteater-api/types';
 
 export interface CourseDetails {
     deptCode: string;
