+++ packages/types/src/index.ts
@@ -7,8 +7,4 @@ export * from './search';
 export * from './websoc';
 export * from './planner';
 export * from './notification';
-export * from '@packages/anteater-api-types/src/calendar';
-export * from '@packages/anteater-api-types/src/courses';
-export * from '@packages/anteater-api-types/src/enrollHist';
-export * from '@packages/anteater-api-types/src/grades';
-export * from '@packages/anteater-api-types/src/websoc';
+export * from '@packages/anteater-api/types';
